refactor(transform): #294 move required-attribute helpers to attribute_required.go (pure move, makes room under the 500-line cap)
